[ switch macro now handles multiple matches ]
This was surprisingly difficult.  The attached tests demonstrate the possible uses.  You can now (switch something ((one two) matches-either-one-or-two))
or (switch something ('(two strings) ...)) or (switch something ('string value)) or (switch something (var multi line body)).

The actual macro is a mess, but the interface feels good, the tests pass and the output js is clean.