[ add foo!, change foo? ;=> fooQ to foo? ;=> foo__QUERY ]
Explanation:

foo-q should not be the same as foo? and foo-b should not be the same as foo!
also, the B in fooB wasn't particularly meaningful, nor was the Q in fooQ.

Slightly more verbose is foo__QUERY and foo__BANG.

I'm planning to use the bang to mean "side effecting" because it would be nice to have a macro shortcut for:

(setf some-string (concat some-string " also this"))

so now I can add
(concat! some-string "string string" var "string, etc")
;=> someString = someString + "string string" + var + "string, etc";