[ add hash and array shortcuts ]
{ a b c d } is a shortcut for (hash a b c d).
[ a b c ] is a shortcut for (list a b c).

I still feel a little weird about it, but it's just a handful of lines.

Technically, { a b c [ d e ) ) would work just as well as { a b c [ d e ] }, but I can't imagine why you'd want to do that.  Not worth going out of my way to make it break, though.

Eventually, I'd like to make this accessible from within the code as simply as (register-shortcut "{" 'hash) or (unregister-shortcut "{").

Since : and , are ignored currently, this means you can do something that looks like json:

{
  a: b,
  c: [ d, e ]
}