[ fix flushing in repl ]
This commit removes the prompt-on-drain and flushed stuff, since it seems that's no longer necessary.  It was initially a hack copied from [1], but it's not in there anymore.

Additionally, and the actual source of the bug [2] was that process.openStdin().write('stuff') wasn't flushing, but process.openStdin().write('stuff\n') does.  No idea why.

[1] https://github.com/ry/node/blob/master/lib/repl.js
[2] https://github.com/jbr/sibilant/issues/#issue/32