[ add if-else macro ]
This provides an easier way of chaining if-else's.

What previously could only have been expressed as

    (if a x (if b y (if c z otherwise)))

can now be expressed as

    (if-else a x
             b y
             c z
             true otherwise)

A possible next step is to handle an odd number of arguments, leaving the last one as a default (so we can skip the ungainly "true" for the fall-through case).