Fixes issue #48, double semi-colons when defvar used within defun
The problem is that defun calls macros.lamba which automatically inserts the semicolon on the end of lines. Since defvar also inserts the semicolon you wind up with 2 semicolons. This isn't a problem in JavaScript, just a prettiness problem.

The fix is to check to see if the last character in the statement is a semicolon, and if it is to not do the insertion.