Process stdin and commandline inputs properly
Importantly, this changes the meaning of -i.  What was previously -i is
now -f, for files.  -i means "wait for standard in."  -e is redefined
as "-x -i" which means sibilant can now translate standard in as well
as executing it.

The options library also was changed to pass all of the options to each
option filter.  This was needed so --input could check if the --execute
flag was set.