Macros return arrays of strings
This is an incremental refactor in the direction of full-fledged quoting
and pseudoquoting, as well as source maps.  As of this commit, the
strategy is to return deeply nested arrays of strings, which are
flattened and concatenated as the very final step.  The next version
will return output nodes that contain a reference to the originating ast
node, and which might be mixed with ast nodes.  When a macro returns ast
nodes instead of output nodes, the returned ast nodes will be evaluated,
allowing any number of (introspectable) intermediate macro evaluation
steps.  This also means that instead of directly calling other macros,
most macros will be defined by returning an ast. This also means that
mixed quoting (with unquoted tokens) can be accomplished by returning
mixed ast nodes and evaluated output nodes.